@@ -21,29 +21,26 @@
     "@angular/platform-browser": "^11.0.0",
     "@angular/platform-browser-dynamic": "^11.0.0",
     "@angular/router": "^11.0.0",
-    "@ionic/angular": "^5.4.0",
-    "@ionic/storage": "2.2.0",
-    "@start9labs/emver": "^0.1.1",
+    "@ionic/angular": "^5.6.0",
+    "@ionic/storage": "^3.0.0",
+    "@ionic/storage-angular": "^3.0.0",
+    "@ngrx/component": "^11.1.1",
+    "@start9labs/emver": "^0.1.4",
     "ajv": "^6.12.6",
-    "angularx-qrcode": "^10.0.11",
-    "base32.js": "^0.1.0",
-    "base64url": "^3.0.1",
-    "bip39": "^3.0.2",
-    "bitcoinjs-lib": "^5.2.0",
+    "angularx-qrcode": "^11.0.0",
     "compare-versions": "^3.5.0",
     "core-js": "^3.4.0",
     "handlebars": "^4.7.6",
     "json-pointer": "^0.6.1",
     "jsonpointerx": "^1.0.30",
-    "jsontokens": "^3.0.0",
     "marked": "^2.0.0",
-    "rxjs": "^6.6.3",
-    "uuid": "^8.3.1",
-    "zone.js": "^0.11.2",
-    "patch-db-client": "file: ../../../../patch-db-client"
+    "patch-db-client": "file: ../../../../patch-db-client",
+    "rxjs": "^6.6.0",
+    "uuid": "^8.3.0",
+    "zone.js": "^0.11.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.0",
+    "@angular-devkit/build-angular": "^0.1102.0",
     "@angular/cli": "^11.0.0",
     "@angular/compiler": "^11.0.0",
     "@angular/compiler-cli": "^11.0.0",
@@ -51,12 +48,12 @@
     "@ionic/angular-toolkit": "^3.0.0",
     "@ionic/lab": "^3.2.9",
     "@types/json-pointer": "^1.0.30",
-    "@types/marked": "^1.1.0",
-    "@types/node": "^14.11.10",
+    "@types/marked": "^2.0.0",
+    "@types/node": "^15.0.0",
     "@types/uuid": "^8.0.0",
-    "node-html-parser": "2.0.0",
+    "node-html-parser": "^3.2.0",
     "ts-node": "^9.1.0",
     "tslint": "^6.1.0",
-    "typescript": "4.0.5"
+    "typescript": "4.1.5"
   }
 }
