@@ -1,8 +1,7 @@
 import { Injectable } from '@angular/core'
-import { AppModel, AppStatus } from 'src/app/models/app-model'
-import { OsUpdateService } from 'src/app/services/os-update.service'
+import { InstalledPackageDataEntry } from 'src/app/models/patch-db/data-model'
+import { Breakages } from 'src/app/services/api/api-types'
 import { exists } from 'src/app/util/misc.util'
-import { AppDependency, DependentBreakage, AppInstalledPreview } from '../../models/app-types'
 import { ApiService } from '../../services/api/api.service'
 import { InstallWizardComponent, SlideDefinition, TopbarParams } from './install-wizard.component'
 import { WizardAction } from './wizard-types'
@@ -11,28 +10,29 @@ import { WizardAction } from './wizard-types'
 export class WizardBaker {
   constructor (
     private readonly apiService: ApiService,
-    private readonly updateService: OsUpdateService,
-    private readonly appModel: AppModel,
   ) { }
 
   install (values: {
-    id: string, title: string, version: string, serviceRequirements: AppDependency[], installAlert?: string
+    id: string, title: string, version: string, installAlert?: string
   }): InstallWizardComponent['params'] {
-    const { id, title, version, serviceRequirements, installAlert } = values
+    const { id, title, version, installAlert } = values
 
     validate(id, exists, 'missing id')
     validate(title, exists, 'missing title')
     validate(version, exists, 'missing version')
-    validate(serviceRequirements, t => !!t && Array.isArray(t), 'missing serviceRequirements')
 
     const action = 'install'
-    const toolbar: TopbarParams  = { action, title, version }
+    const toolbar: TopbarParams = { action, title, version }
 
     const slideDefinitions: SlideDefinition[] = [
       installAlert ? {
         slide: {
           selector: 'notes',
-          params: { notes: installAlert, title: 'Warning', titleColor: 'warning' },
+          params: {
+            notes: installAlert,
+            title: 'Warning',
+            titleColor: 'warning',
+          },
         },
         bottomBar: {
           cancel: { afterLoading: { text: 'Cancel' } }, next: 'Next',
@@ -40,22 +40,12 @@ export class WizardBaker {
       } : undefined,
       {
         slide: {
-          selector: 'dependencies',
-          params: { action, title, version, serviceRequirements },
-        },
-        bottomBar: {
-          cancel: { afterLoading: { text: 'Cancel' } },
-          next: 'Install',
-        },
-      },
-      {
-        slide: {
           selector: 'complete',
           params: {
             action,
             verb: 'beginning installation for',
             title,
-            executeAction: () => this.apiService.installApp(id, version).then(app => { this.appModel.add({ ...app, status: AppStatus.INSTALLING })}),
+            executeAction: () => this.apiService.installPackage({ id, version }),
           },
         },
         bottomBar: {
@@ -68,14 +58,13 @@ export class WizardBaker {
   }
 
   update (values: {
-    id: string, title: string, version: string, serviceRequirements: AppDependency[], installAlert?: string
+    id: string, title: string, version: string, installAlert?: string
   }): InstallWizardComponent['params'] {
-    const { id, title, version, serviceRequirements, installAlert } = values
+    const { id, title, version, installAlert } = values
 
     validate(id, exists, 'missing id')
     validate(title, exists, 'missing title')
     validate(version, exists, 'missing version')
-    validate(serviceRequirements, t => !!t && Array.isArray(t), 'missing serviceRequirements')
 
     const action = 'update'
     const toolbar: TopbarParams  = { action, title, version }
@@ -84,26 +73,26 @@ export class WizardBaker {
       installAlert ? {
         slide: {
           selector: 'notes',
-          params: { notes: installAlert, title: 'Warning', titleColor: 'warning'},
+          params: {
+            notes: installAlert,
+            title: 'Warning',
+            titleColor: 'warning',
+          },
         },
         bottomBar: {
           cancel: { afterLoading: { text: 'Cancel' } },
           next: 'Next',
         },
       } : undefined,
-      { slide: {
-          selector: 'dependencies',
-          params: { action, title, version, serviceRequirements },
-        },
-        bottomBar: {
-          cancel: { afterLoading: { text: 'Cancel' } },
-          next: 'Update',
-        },
-      },
-      { slide: {
+      {
+        slide: {
           selector: 'dependents',
           params: {
-            skipConfirmationDialogue: true, action, verb: 'updating', title, fetchBreakages: () => this.apiService.installApp(id, version, true).then( ({ breakages }) => breakages ),
+            skipConfirmationDialogue: true,
+            action,
+            verb: 'updating',
+            title,
+            fetchBreakages: () => this.apiService.dryUpdatePackage({ id, version }).then( ({ breakages }) => breakages ),
           },
         },
         bottomBar: {
@@ -111,12 +100,14 @@ export class WizardBaker {
           next: 'Update Anyways',
         },
       },
-      { slide: {
+      {
+        slide: {
           selector: 'complete',
           params: {
-            action, verb: 'beginning update for', title, executeAction: () => this.apiService.installApp(id, version).then(app => {
-              this.appModel.update({ id: app.id, status: AppStatus.INSTALLING })
-            }),
+            action,
+            verb: 'beginning update for',
+            title,
+            executeAction: () => this.apiService.installPackage({ id, version }),
           },
         },
         bottomBar: {
@@ -138,18 +129,27 @@ export class WizardBaker {
     const toolbar: TopbarParams  = { action, title, version }
 
     const slideDefinitions: SlideDefinition[] = [
-      { slide : {
+      {
+        slide : {
           selector: 'notes',
-          params: { notes: releaseNotes, title: 'Release Notes', titleColor: 'dark' },
+          params: {
+            notes: releaseNotes,
+            title: 'Release Notes',
+            titleColor: 'dark',
+          },
         },
         bottomBar: {
           cancel: { afterLoading: { text: 'Cancel' } }, next: 'Update OS',
         },
       },
-      { slide: {
+      {
+        slide: {
           selector: 'complete',
           params: {
-            action, verb: 'beginning update for', title, executeAction: () => this.updateService.updateEmbassyOS(version),
+            action,
+            verb: 'beginning update for',
+            title,
+            executeAction: () => this.apiService.updateServer({ }),
           },
         },
         bottomBar: {
@@ -162,14 +162,13 @@ export class WizardBaker {
   }
 
   downgrade (values: {
-    id: string, title: string, version: string, serviceRequirements: AppDependency[], installAlert?: string
+    id: string, title: string, version: string, installAlert?: string
   }): InstallWizardComponent['params'] {
-    const { id, title, version, serviceRequirements, installAlert } = values
+    const { id, title, version, installAlert } = values
 
     validate(id, exists, 'missing id')
     validate(title, exists, 'missing title')
     validate(version, exists, 'missing version')
-    validate(serviceRequirements, t => !!t && Array.isArray(t), 'missing serviceRequirements')
 
     const action = 'downgrade'
     const toolbar: TopbarParams  = { action, title, version }
@@ -178,23 +177,22 @@ export class WizardBaker {
       installAlert ? {
         slide: {
           selector: 'notes',
-          params: { notes: installAlert, title: 'Warning', titleColor: 'warning' },
+          params: {
+            notes: installAlert,
+            title: 'Warning',
+            titleColor: 'warning',
+          },
         },
         bottomBar: { cancel: { afterLoading: { text: 'Cancel' } }, next: 'Next' },
       } : undefined,
       { slide: {
-          selector: 'dependencies',
-          params: { action, title, version, serviceRequirements },
-        },
-        bottomBar: {
-          cancel: { afterLoading: { text: 'Cancel' } },
-          next: 'Downgrade',
-        },
-      },
-      { slide: {
           selector: 'dependents',
           params: {
-            skipConfirmationDialogue: true, action, verb: 'downgrading', title, fetchBreakages: () => this.apiService.installApp(id, version, true).then( ({ breakages }) => breakages ),
+            skipConfirmationDialogue: true,
+            action,
+            verb: 'downgrading',
+            title,
+            fetchBreakages: () => this.apiService.dryUpdatePackage({ id, version }).then( ({ breakages }) => breakages ),
           },
         },
         bottomBar: {
@@ -204,9 +202,10 @@ export class WizardBaker {
       { slide: {
           selector: 'complete',
           params: {
-            action, verb: 'beginning downgrade for', title, executeAction: () => this.apiService.installApp(id, version).then(app => {
-              this.appModel.update({ id: app.id, status: AppStatus.INSTALLING })
-            }),
+            action,
+            verb: 'beginning downgrade for',
+            title,
+            executeAction: () => this.apiService.installPackage({ id, version }),
           },
         },
         bottomBar: {
@@ -231,7 +230,8 @@ export class WizardBaker {
     const toolbar: TopbarParams  = { action, title, version }
 
     const slideDefinitions: SlideDefinition[] = [
-      { slide: {
+      {
+        slide: {
           selector: 'notes',
           params: {
             notes: uninstallAlert || defaultUninstallationWarning(title),
@@ -241,18 +241,26 @@ export class WizardBaker {
         },
         bottomBar: { cancel: { afterLoading: { text: 'Cancel' } }, next: 'Continue' },
       },
-      { slide: {
+      {
+        slide: {
           selector: 'dependents',
           params: {
-            action, verb: 'uninstalling', title, fetchBreakages: () => this.apiService.uninstallApp(id, true).then( ({ breakages }) => breakages ),
+            action,
+            verb: 'uninstalling',
+            title,
+            fetchBreakages: () => this.apiService.dryRemovePackage({ id }).then( ({ breakages }) => breakages ),
           },
         },
         bottomBar: { cancel: { whileLoading: { }, afterLoading: { text: 'Cancel' } }, next: 'Uninstall' },
       },
-      { slide: {
+      {
+        slide: {
           selector: 'complete',
           params: {
-            action, verb: 'uninstalling', title, executeAction: () => this.apiService.uninstallApp(id).then(() => this.appModel.delete(id)),
+            action,
+            verb: 'uninstalling',
+            title,
+            executeAction: () => this.apiService.removePackage({ id }),
           },
         },
         bottomBar: { finish: 'Dismiss', cancel: { whileLoading: { } } },
@@ -262,7 +270,7 @@ export class WizardBaker {
   }
 
   stop (values: {
-    breakages: DependentBreakage[], id: string, title: string, version: string
+    breakages: Breakages, id: string, title: string, version: string
   }): InstallWizardComponent['params'] {
     const { breakages, title, version } = values
 
@@ -274,10 +282,14 @@ export class WizardBaker {
     const toolbar: TopbarParams  = { action, title, version }
 
     const slideDefinitions: SlideDefinition[] = [
-      { slide: {
+      {
+        slide: {
           selector: 'dependents',
           params: {
-            action, verb: 'stopping', title, fetchBreakages: () => Promise.resolve(breakages),
+            action,
+            verb: 'stopping',
+            title,
+            fetchBreakages: () => Promise.resolve(breakages),
           },
         },
         bottomBar: { cancel: { afterLoading: { text: 'Cancel' } }, next: 'Stop Anyways' },
@@ -286,19 +298,20 @@ export class WizardBaker {
     return { toolbar, slideDefinitions }
   }
 
-  configure (values: {
-    breakages: DependentBreakage[], app: AppInstalledPreview
-  }): InstallWizardComponent['params'] {
-    const { breakages, app } = values
-    const { title, versionInstalled: version  } = app
+  configure (values: { breakages: Breakages, pkg: InstalledPackageDataEntry }): InstallWizardComponent['params'] {
+    const { breakages, pkg } = values
+    const { title, version } = pkg.manifest
     const action = 'configure'
     const toolbar: TopbarParams  = { action, title, version }
 
     const slideDefinitions: SlideDefinition[] = [
-      { slide: {
+      {
+        slide: {
           selector: 'dependents',
           params: {
-            action, verb: 'saving config for', title, fetchBreakages: () => Promise.resolve(breakages),
+            action,
+            verb: 'saving config for',
+            title, fetchBreakages: () => Promise.resolve(breakages),
           },
         },
         bottomBar: { cancel: { afterLoading: { text: 'Cancel' } }, next: 'Save Config Anyways' },
