@@ -1,81 +1,64 @@
-import { Component, NgZone } from '@angular/core'
+import { Component } from '@angular/core'
 import { ActivatedRoute } from '@angular/router'
-import { AppAvailableFull, AppAvailableVersionSpecificInfo } from 'src/app/models/app-types'
 import { ApiService } from 'src/app/services/api/api.service'
 import { AlertController, ModalController, NavController, PopoverController } from '@ionic/angular'
-import { markAsLoadingDuring$ } from 'src/app/services/loader.service'
-import { BehaviorSubject, from, Observable, of } from 'rxjs'
-import { catchError, concatMap, filter, switchMap, tap } from 'rxjs/operators'
 import { Recommendation } from 'src/app/components/recommendation-button/recommendation-button.component'
 import { wizardModal } from 'src/app/components/install-wizard/install-wizard.component'
 import { WizardBaker } from 'src/app/components/install-wizard/prebaked-wizards'
-import { AppModel } from 'src/app/models/app-model'
-import { initPropertySubject, peekProperties, PropertySubject } from 'src/app/util/property-subject.util'
-import { Cleanup } from 'src/app/util/cleanup'
 import { InformationPopoverComponent } from 'src/app/components/information-popover/information-popover.component'
 import { Emver } from 'src/app/services/emver.service'
 import { displayEmver } from 'src/app/pipes/emver.pipe'
 import { pauseFor } from 'src/app/util/misc.util'
+import { AvailableShow } from 'src/app/services/api/api-types'
+import { PatchDbModel } from 'src/app/models/patch-db/patch-db-model'
+import { PackageState } from 'src/app/models/patch-db/data-model'
 
 @Component({
   selector: 'app-available-show',
   templateUrl: './app-available-show.page.html',
   styleUrls: ['./app-available-show.page.scss'],
 })
-export class AppAvailableShowPage extends Cleanup {
-  $loading$ = new BehaviorSubject(true)
+export class AppAvailableShowPage {
+  loading = true
+  error = ''
+  pkg: AvailableShow
+  pkgId: string
 
-  // When a new version is selected
-  $newVersionLoading$ = new BehaviorSubject(false)
-  // When dependencies are refreshing
-  $dependenciesLoading$ = new BehaviorSubject(false)
+  PackageState = PackageState
 
-  $error$ = new BehaviorSubject(undefined)
-  $app$: PropertySubject<AppAvailableFull> = { } as any
-  appId: string
+  rec: Recommendation | null = null
+  showRec = true
 
-  openRecommendation = false
-  recommendation: Recommendation | null = null
-
-  serviceDependencyDefintion = '<span style="font-style: italic">Service Dependencies</span> are other services that this service recommends or requires in order to run.'
+  depDefinition = '<span style="font-style: italic">Service Dependencies</span> are other services that this service recommends or requires in order to run.'
 
   constructor (
     private readonly route: ActivatedRoute,
     private readonly apiService: ApiService,
     private readonly alertCtrl: AlertController,
-    private readonly zone: NgZone,
     private readonly modalCtrl: ModalController,
     private readonly wizardBaker: WizardBaker,
     private readonly navCtrl: NavController,
-    private readonly appModel: AppModel,
     private readonly popoverController: PopoverController,
     private readonly emver: Emver,
-  ) {
-    super()
-  }
+    public readonly patch: PatchDbModel,
+  ) { }
 
   async ngOnInit () {
-    this.appId = this.route.snapshot.paramMap.get('appId') as string
-
-    this.cleanup(
-      // new version always includes dependencies, but not vice versa
-      this.$newVersionLoading$.subscribe(this.$dependenciesLoading$),
-      markAsLoadingDuring$(this.$loading$,
-        from(this.apiService.getAvailableApp(this.appId)).pipe(
-          tap(app => this.$app$ = initPropertySubject(app)),
-          concatMap(() => this.fetchRecommendation()),
-        ),
-      ).pipe(
-        concatMap(() => this.syncWhenDependencyInstalls()), //must be final in stack
-        catchError(e => of(this.setError(e))),
-      ).subscribe(),
-    )
+    this.pkgId = this.route.snapshot.paramMap.get('pkgId') as string
+    this.rec = history.state && history.state.installRec as Recommendation
+    this.getPkg()
   }
 
-  ionViewDidEnter () {
-    markAsLoadingDuring$(this.$dependenciesLoading$, this.syncVersionSpecificInfo()).subscribe({
-      error: e => this.setError(e),
-    })
+  async getPkg (version?: string): Promise<void> {
+    this.loading = true
+    try {
+      this.pkg = await this.apiService.getAvailableShow({ id: this.pkgId, version })
+    } catch (e) {
+      console.error(e)
+      this.error = e.message
+    } finally {
+      this.loading = false
+    }
   }
 
   async presentPopover (information: string, ev: any) {
@@ -92,34 +75,17 @@ export class AppAvailableShowPage extends Cleanup {
     return await popover.present()
   }
 
-  syncVersionSpecificInfo (versionSpec?: string): Observable<any> {
-    if (!this.$app$.versionViewing) return of({ })
-    const specToFetch = versionSpec || `=${this.$app$.versionViewing.getValue()}`
-    return from(this.apiService.getAvailableAppVersionSpecificInfo(this.appId, specToFetch)).pipe(
-      tap(versionInfo => this.mergeInfo(versionInfo)),
-    )
-  }
-
-  private mergeInfo (versionSpecificInfo: AppAvailableVersionSpecificInfo) {
-    this.zone.run(() => {
-      Object.entries(versionSpecificInfo).forEach( ([k, v]) => {
-        if (!this.$app$[k]) this.$app$[k] = new BehaviorSubject(undefined)
-        if (v !== this.$app$[k].getValue()) this.$app$[k].next(v)
-      })
-    })
-  }
-
   async presentAlertVersions () {
-    const app = peekProperties(this.$app$)
     const alert = await this.alertCtrl.create({
       header: 'Versions',
       backdropDismiss: false,
-      inputs: app.versions.sort((a, b) => -1 * this.emver.compare(a, b)).map(v => {
-        return { name: v, // for CSS
+      inputs: this.pkg.versions.sort((a, b) => -1 * this.emver.compare(a, b)).map(v => {
+        return {
+          name: v, // for CSS
           type: 'radio',
           label: displayEmver(v), // appearance on screen
           value: v, // literal SEM version value
-          checked: app.versionViewing === v,
+          checked: this.pkg.manifest.version === v,
         }
       }),
       buttons: [
@@ -129,17 +95,7 @@ export class AppAvailableShowPage extends Cleanup {
         }, {
           text: 'Ok',
           handler: (version: string) => {
-            const previousVersion = this.$app$.versionViewing.getValue()
-            this.$app$.versionViewing.next(version)
-            markAsLoadingDuring$(
-              this.$newVersionLoading$, this.syncVersionSpecificInfo(`=${version}`),
-            )
-            .subscribe({
-              error: e => {
-                this.setError(e)
-                this.$app$.versionViewing.next(previousVersion)
-              },
-            })
+            this.getPkg(version)
           },
         },
       ],
@@ -149,15 +105,14 @@ export class AppAvailableShowPage extends Cleanup {
   }
 
   async install () {
-    const app = peekProperties(this.$app$)
+    const { id, title, version, dependencies, alerts } = this.pkg.manifest
     const { cancelled } = await wizardModal(
       this.modalCtrl,
       this.wizardBaker.install({
-        id: app.id,
-        title: app.title,
-        version: app.versionViewing,
-        serviceRequirements: app.serviceRequirements,
-        installAlert: app.installAlert,
+        id,
+        title,
+        version,
+        installAlert: alerts.install,
       }),
     )
     if (cancelled) return
@@ -166,14 +121,13 @@ export class AppAvailableShowPage extends Cleanup {
   }
 
   async update (action: 'update' | 'downgrade') {
-    const app = peekProperties(this.$app$)
-
+    const { id, title, version, dependencies, alerts } = this.pkg.manifest
     const value = {
-      id: app.id,
-      title: app.title,
-      version: app.versionViewing,
-      serviceRequirements: app.serviceRequirements,
-      installAlert: app.installAlert,
+      id,
+      title,
+      version,
+      serviceRequirements: dependencies,
+      installAlert: alerts.install,
     }
 
     const { cancelled } = await wizardModal(
@@ -188,27 +142,7 @@ export class AppAvailableShowPage extends Cleanup {
     this.navCtrl.back()
   }
 
-  private fetchRecommendation (): Observable<any> {
-    this.recommendation = history.state && history.state.installationRecommendation
-
-    if (this.recommendation) {
-      return from(this.syncVersionSpecificInfo(this.recommendation.versionSpec))
-    } else {
-      return of({ })
-    }
-  }
-
-  private syncWhenDependencyInstalls (): Observable<void> {
-    return this.$app$.serviceRequirements.pipe(
-      filter(deps => !!deps),
-      switchMap(deps => this.appModel.watchForInstallations(deps)),
-      concatMap(() => markAsLoadingDuring$(this.$dependenciesLoading$, this.syncVersionSpecificInfo())),
-      catchError(e => of(console.error(e))),
-    )
-  }
-
-  private setError (e: Error) {
-    console.error(e)
-    this.$error$.next(e.message)
+  dismissRec () {
+    this.showRec = false
   }
 }
