@@ -1,80 +1,73 @@
 import { Component } from '@angular/core'
 import { NavController, AlertController, ModalController, PopoverController } from '@ionic/angular'
 import { ActivatedRoute } from '@angular/router'
-import { AppModel, AppStatus } from 'src/app/models/app-model'
-import { AppInstalledFull } from 'src/app/models/app-types'
 import { ApiService } from 'src/app/services/api/api.service'
-import { pauseFor, isEmptyObject, modulateTime } from 'src/app/util/misc.util'
-import { PropertySubject, peekProperties } from 'src/app/util/property-subject.util'
-import { LoaderService, markAsLoadingDuring$ } from 'src/app/services/loader.service'
+import { isEmptyObject } from 'src/app/util/misc.util'
+import { LoaderService } from 'src/app/services/loader.service'
 import { TrackingModalController } from 'src/app/services/tracking-modal-controller.service'
-import { ModelPreload } from 'src/app/models/model-preload'
-import { BehaviorSubject, forkJoin, from, fromEvent, of } from 'rxjs'
+import { BehaviorSubject, from, fromEvent, of, Subscription } from 'rxjs'
 import { catchError, concatMap, map, take, tap } from 'rxjs/operators'
 import { Recommendation } from 'src/app/components/recommendation-button/recommendation-button.component'
 import { wizardModal } from 'src/app/components/install-wizard/install-wizard.component'
 import { WizardBaker } from 'src/app/components/install-wizard/prebaked-wizards'
-import { Cleanup } from 'src/app/util/cleanup'
 import { InformationPopoverComponent } from 'src/app/components/information-popover/information-popover.component'
-import { ConfigSpec } from 'src/app/app-config/config-types'
-import { ConfigCursor } from 'src/app/app-config/config-cursor'
+import { ConfigSpec } from 'src/app/pkg-config/config-types'
+import { ConfigCursor } from 'src/app/pkg-config/config-cursor'
+import { InstalledPackageDataEntry } from 'src/app/models/patch-db/data-model'
+import { PatchDbModel } from 'src/app/models/patch-db/patch-db-model'
 
 @Component({
   selector: 'app-config',
   templateUrl: './app-config.page.html',
   styleUrls: ['./app-config.page.scss'],
 })
-export class AppConfigPage extends Cleanup {
+export class AppConfigPage {
   error: { text: string, moreInfo?:
     { title: string, description: string, buttonText: string }
   }
 
-  invalid: string
-  $loading$ = new BehaviorSubject(true)
-  $loadingText$ = new BehaviorSubject(undefined)
+  loadingText$ = new BehaviorSubject(undefined)
 
-  app: PropertySubject<AppInstalledFull> = { } as any
-  appId: string
+  pkg: InstalledPackageDataEntry
   hasConfig = false
 
-  recommendation: Recommendation | null = null
-  showRecommendation = true
-  openRecommendation = false
+  backButtonDefense = false
 
+  rec: Recommendation | null = null
+  showRec = true
+  openRec = false
+
+  invalid: string
   edited: boolean
   added: boolean
   rootCursor: ConfigCursor<'object'>
   spec: ConfigSpec
   config: object
 
-  AppStatus = AppStatus
+  subs: Subscription[]
 
   constructor (
     private readonly navCtrl: NavController,
     private readonly route: ActivatedRoute,
     private readonly wizardBaker: WizardBaker,
-    private readonly preload: ModelPreload,
     private readonly apiService: ApiService,
     private readonly loader: LoaderService,
     private readonly alertCtrl: AlertController,
     private readonly modalController: ModalController,
     private readonly trackingModalCtrl: TrackingModalController,
     private readonly popoverController: PopoverController,
-    private readonly appModel: AppModel,
-  ) { super() }
-
-  backButtonDefense = false
+    private readonly patch: PatchDbModel,
+  ) { }
 
   async ngOnInit () {
-    this.appId = this.route.snapshot.paramMap.get('appId') as string
+    const pkgId = this.route.snapshot.paramMap.get('pkgId') as string
 
-    this.route.params.pipe(take(1)).subscribe(params => {
-      if (params.edit) {
-        window.history.back()
-      }
-    })
-
-    this.cleanup(
+    this.subs = [
+      this.route.params.pipe(take(1)).subscribe(params => {
+        if (params.edit) {
+          window.history.back()
+        }
+      }),
       fromEvent(window, 'popstate').subscribe(() => {
         this.backButtonDefense = false
         this.trackingModalCtrl.dismissAll()
@@ -90,49 +83,51 @@ export class AppConfigPage extends Cleanup {
           this.navCtrl.back()
         }
       }),
-    )
+    ]
 
-    markAsLoadingDuring$(this.$loading$,
-      from(this.preload.appFull(this.appId))
-        .pipe(
-          tap(app => this.app = app),
-          tap(() => this.$loadingText$.next(`Fetching config spec...`)),
-          concatMap(() => forkJoin([this.apiService.getAppConfig(this.appId), pauseFor(600)])),
-          concatMap(([{ spec, config }]) => {
-            const rec = history.state && history.state.configRecommendation as Recommendation
-            if (rec) {
-              this.$loadingText$.next(`Setting properties to accomodate ${rec.title}...`)
-              return from(this.apiService.postConfigureDependency(this.appId, rec.appId, true))
-              .pipe(
-                map(res => ({
-                  spec,
-                  config,
-                  dependencyConfig: res.config,
-                })),
-                tap(() => this.recommendation = rec),
-                catchError(e => {
-                  this.error = { text: `Could not set properties to accomodate ${rec.title}: ${e.message}`, moreInfo: {
-                    title: `${rec.title} requires the following:`,
-                    description: rec.description,
-                    buttonText: 'Configure Manually',
-                  } }
-                  return of({ spec, config, dependencyConfig: null })
-                }),
-              )
-            } else {
+    this.patch.watch$('package-data', pkgId, 'installed')
+    .pipe(
+      tap(pkg => this.pkg = pkg),
+      tap(() => this.loadingText$.next(`Fetching config spec...`)),
+      concatMap(() => this.apiService.getPackageConfig({ id: pkgId })),
+      concatMap(({ spec, config }) => {
+        const rec = history.state && history.state.configRecommendation as Recommendation
+        if (rec) {
+          this.loadingText$.next(`Setting properties to accommodate ${rec.dependentTitle}...`)
+          return from(this.apiService.dryConfigureDependency({ 'dependency-id': pkgId, 'dependent-id': rec.dependentId }))
+          .pipe(
+            map(res => ({
+              spec,
+              config,
+              dependencyConfig: res,
+            })),
+            tap(() => this.rec = rec),
+            catchError(e => {
+              this.error = { text: `Could not set properties to accommodate ${rec.dependentTitle}: ${e.message}`, moreInfo: {
+                title: `${rec.dependentTitle} requires the following:`,
+                description: rec.description,
+                buttonText: 'Configure Manually',
+              } }
               return of({ spec, config, dependencyConfig: null })
-            }
-          }),
-          map(({ spec, config, dependencyConfig }) => this.setConfig(spec, config, dependencyConfig)),
-          tap(() => this.$loadingText$.next(undefined)),
-        ),
+            }),
+          )
+        } else {
+          return of({ spec, config, dependencyConfig: null })
+        }
+      }),
+      map(({ spec, config, dependencyConfig }) => this.setConfig(spec, config, dependencyConfig)),
+      tap(() => this.loadingText$.next(undefined)),
+      take(1),
     ).subscribe({
-        error: e => {
-          console.error(e)
-          this.error = { text: e.message }
-        },
+      error: e => {
+        console.error(e.message)
+        this.error = { text: e.message }
       },
-    )
+    })
+  }
+
+  ngOnDestroy () {
+    this.subs.forEach(sub => sub.unsubscribe())
   }
 
   async presentPopover (title: string, description: string, ev: any) {
@@ -165,8 +160,8 @@ export class AppConfigPage extends Cleanup {
     this.hasConfig = !isEmptyObject(this.spec)
   }
 
-  dismissRecommendation () {
-    this.showRecommendation = false
+  dismissRec () {
+    this.showRec = false
   }
 
   dismissError () {
@@ -181,38 +176,30 @@ export class AppConfigPage extends Cleanup {
     }
   }
 
-  async save () {
-    const app = peekProperties(this.app)
-    const ogAppStatus = app.status
-
+  async save (pkg: InstalledPackageDataEntry) {
     return this.loader.of({
       message: `Saving config...`,
       spinner: 'lines',
       cssClass: 'loader',
     }).displayDuringAsync(async () => {
-      const config = this.config
-      const { breakages } = await this.apiService.patchAppConfig(app, config, true)
+      const { breakages } = await this.apiService.drySetPackageConfig({ id: pkg.manifest.id, config: this.config })
 
       if (breakages.length) {
         const { cancelled } = await wizardModal(
           this.modalController,
           this.wizardBaker.configure({
-            app,
+            pkg,
             breakages,
           }),
         )
         if (cancelled) return { skip: true }
       }
 
-      return this.apiService.patchAppConfig(app, config).then(
-        () => this.preload.loadInstalledApp(this.appId).then(() => ({ skip: false })),
-      )
+      return this.apiService.setPackageConfig({ id: pkg.manifest.id, config: this.config })
+        .then(() => ({ skip: false }))
     })
     .then(({ skip }) => {
       if (skip) return
-      if (ogAppStatus === AppStatus.RUNNING) {
-        this.appModel.update({ id: this.appId, status: AppStatus.RESTARTING }, modulateTime(new Date(), 3, 'seconds'))
-      }
       this.navCtrl.back()
     })
     .catch(e => this.error = { text: e.message })
