@@ -5,6 +5,7 @@ import { DevOptionsPage } from './dev-options.page'
 import { Routes, RouterModule } from '@angular/router'
 import { PwaBackComponentModule } from 'src/app/components/pwa-back-button/pwa-back.component.module'
 import { ObjectConfigComponentModule } from 'src/app/components/object-config/object-config.component.module'
+import { SharingModule } from 'src/app/modules/sharing.module'
 
 const routes: Routes = [
   {
@@ -20,7 +21,10 @@ const routes: Routes = [
     ObjectConfigComponentModule,
     RouterModule.forChild(routes),
     PwaBackComponentModule,
+    SharingModule,
   ],
-  declarations: [DevOptionsPage],
+  declarations: [
+    DevOptionsPage,
+  ],
 })
 export class DevOptionsPageModule { }
