@@ -1,12 +1,8 @@
 import { Component } from '@angular/core'
-import { SSHFingerprint, S9Server } from 'src/app/models/server-model'
-import { ApiService } from 'src/app/services/api/api.service'
-import { pauseFor } from 'src/app/util/misc.util'
-import { PropertySubject } from 'src/app/util/property-subject.util'
 import { ServerConfigService } from 'src/app/services/server-config.service'
-import { LoaderService } from 'src/app/services/loader.service'
-import { ModelPreload } from 'src/app/models/model-preload'
 import { AlertController } from '@ionic/angular'
+import { LoaderService } from 'src/app/services/loader.service'
+import { SSHService } from './ssh.service'
 
 @Component({
   selector: 'dev-ssh-keys',
@@ -14,40 +10,31 @@ import { AlertController } from '@ionic/angular'
   styleUrls: ['dev-ssh-keys.page.scss'],
 })
 export class DevSSHKeysPage {
-  server: PropertySubject<S9Server> = { } as any
   error = ''
+  loading = true
 
   constructor (
-    private readonly apiService: ApiService,
     private readonly loader: LoaderService,
-    private readonly preload: ModelPreload,
     private readonly serverConfigService: ServerConfigService,
     private readonly alertCtrl: AlertController,
+    public readonly sshService: SSHService,
   ) { }
 
   ngOnInit () {
-    this.loader.displayDuring$(
-      this.preload.server(),
-    ).subscribe(s => this.server = s)
-  }
-
-  async doRefresh (event: any) {
-    await Promise.all([
-      this.apiService.getServer(),
-      pauseFor(600),
-    ])
-    event.target.complete()
+    this.sshService.getKeys().then(() => {
+      this.loading = false
+    })
   }
 
   async presentModalAdd () {
-    await this.serverConfigService.presentModalValueEdit('ssh', true)
+    await this.serverConfigService.presentModalValueEdit('ssh')
   }
 
-  async presentAlertDelete (fingerprint: SSHFingerprint) {    
+  async presentAlertDelete (hash: string) {
     const alert = await this.alertCtrl.create({
       backdropDismiss: false,
       header: 'Caution',
-      message: `Are you sure you want to delete this SSH key?`,
+      message: `Are you sure you want to delete this key?`,
       buttons: [
         {
           text: 'Cancel',
@@ -57,7 +44,7 @@ export class DevSSHKeysPage {
           text: 'Delete',
           cssClass: 'alert-danger',
           handler: () => {
-            this.delete(fingerprint)
+            this.delete(hash)
           },
         },
       ],
@@ -65,16 +52,21 @@ export class DevSSHKeysPage {
     await alert.present()
   }
 
-  async delete (fingerprint: SSHFingerprint) {
+  async delete (hash: string): Promise<void> {
+    this.error = ''
     this.loader.of({
       message: 'Deleting...',
       spinner: 'lines',
       cssClass: 'loader',
-    }).displayDuringP(
-      this.apiService.deleteSSHKey(fingerprint).then(() => this.error = ''),
-    ).catch(e => {
+    }).displayDuringAsync(async () => {
+      await this.sshService.delete(hash)
+    }).catch(e => {
       console.error(e)
-      this.error = e.message
+      this.error = ''
     })
+  }
+
+  asIsOrder (a: any, b: any) {
+    return 0
   }
 }
