@@ -1,9 +1,6 @@
 import { Component, ViewChild } from '@angular/core'
 import { ApiService } from 'src/app/services/api/api.service'
 import { IonContent } from '@ionic/angular'
-import { pauseFor } from 'src/app/util/misc.util'
-import { markAsLoadingDuringP } from 'src/app/services/loader.service'
-import { BehaviorSubject } from 'rxjs'
 
 @Component({
   selector: 'server-logs',
@@ -12,7 +9,7 @@ import { BehaviorSubject } from 'rxjs'
 })
 export class ServerLogsPage {
   @ViewChild(IonContent, { static: false }) private content: IonContent
-  $loading$ = new BehaviorSubject(true)
+  loading = true
   error = ''
   logs: string
 
@@ -20,25 +17,23 @@ export class ServerLogsPage {
     private readonly apiService: ApiService,
   ) { }
 
-  async ngOnInit () {
-    markAsLoadingDuringP(this.$loading$, Promise.all([
-      this.getLogs(),
-      pauseFor(600),
-    ]))
+  ngOnInit () {
+    this.getLogs()
   }
 
   async getLogs () {
     this.logs = ''
-    this.$loading$.next(true)
+    this.loading = true
     try {
-      this.logs = (await this.apiService.getServerLogs()).join('\n')
+      const logs = await this.apiService.getServerLogs({ })
+      this.logs = logs.map(l => `${l.timestamp} ${l.log}`).join('\n\n')
       this.error = ''
       setTimeout(async () => await this.content.scrollToBottom(100), 200)
     } catch (e) {
       console.error(e)
       this.error = e.message
     } finally {
-      this.$loading$.next(false)
+      this.loading = false
     }
   }
 }
