@@ -1,48 +1,43 @@
 import { NgModule } from '@angular/core'
 import { Routes, RouterModule } from '@angular/router'
-import { AuthGuard } from '../../guards/auth.guard'
 
 const routes: Routes = [
   {
     path: '',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./server-show/server-show.module').then(m => m.ServerShowPageModule),
   },
   {
+    path: 'backup',
+    loadChildren: () => import('./server-backup/server-backup.module').then(m => m.ServerBackupPageModule),
+  },
+  {
     path: 'specs',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./server-specs/server-specs.module').then(m => m.ServerSpecsPageModule),
   },
   {
     path: 'metrics',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./server-metrics/server-metrics.module').then(m => m.ServerMetricsPageModule),
   },
   {
     path: 'logs',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./server-logs/server-logs.module').then(m => m.ServerLogsPageModule),
   },
   {
-    path: 'config',
-    canActivate: [AuthGuard],
-    loadChildren: () => import('./server-config/server-config.module').then(m => m.ServerConfigPageModule),
+    path: 'settings',
+    loadChildren: () => import('./general-settings/general-settings.module').then(m => m.GeneralSettingsPageModule),
   },
   {
     path: 'wifi',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./wifi/wifi.module').then(m => m.WifiListPageModule),
   },
   {
     path: 'lan',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./lan/lan.module').then(m => m.LANPageModule),
   },
   {
     path: 'developer',
-    canActivate: [AuthGuard],
     loadChildren: () => import('./developer-routes/developer-routing.module').then( m => m.DeveloperRoutingModule),
-  }
+  },
 ]
 
 @NgModule({
